converted rasa files to v2.x
