+++ rasa/config.yml
@@ -24,5 +24,3 @@ policies:
   - name: TEDPolicy
     max_history: 5
     epochs: 100
-  - name: MappingPolicy
-  - name: FormPolicy
