--- rasa/data/nlu.md
@@ -1,177 +1,0 @@
-## intent:start
-- /start
-
-## intent:greet
-- hey
-- hallo
-- hi
-- guten morgen
-- guten abend
-- guten mittag
-- servus
-- hello
-
-## intent:how_are_you
-- wie geht es dir?
-- dir?
-- wie fühlst du dich?
-- wie ist es bei dir?
-- wie geht es dir?
-- wie geht's dir?
-- wie gehts dir?
-- wie ists bei dir?
-- bei dir?
-
-## intent:goodbye
-- bye
-- ciao
-- tschüss
-- bis später
-- bis dann
-- auf wiedersehen
-- mach's gut
-- mach es gut
-- bis bald
-- wiedersehen
-
-## intent:affirm
-- yes
-- ja
-- tatsächlich
-- genau
-- natürlich
-- korrekt
-- klingt gut
-- richtig
-- absolut richtig
-- ganz genau
-- absolut
-- stimmt
-
-## intent:deny
-- no
-- nein
-- nie
-- niemals
-- never ever
-- glaube ich nicht
-- gefällt mir nicht
-- no way
-- nicht wirklich
-- ne
-
-## intent:mood_great
-- gut
-- perfekt
-- mir geht es super
-- fühle mich super
-- mir gehts klasse
-- very good
-- fühl mich sehr gut
-- hervorragend
-- wunderbar
-- alles prima
-- alles paletti
-- mir geht es sehr gut
-- ich fühle mich sehr gut
-- großartig
-- mir geht es gut
-- perfect
-- happy
-- es geht mir gut
-
-## intent:mood_unhappy
-- sad
-- unhappy
-- traurig
-- sehr traurig
-- unglücklich
-- bad
-- very bad
-- schlecht
-- sehr schlecht
-- schrecklich
-- nicht sehr gut
-- so schade
-- nicht gut
-- es geht mir schlecht
-- mir geht es schlecht
-- ich fühle mich nicht so gut
-- ich bin traurig
-- gestresst
-- angespannt
-- ich bin traurig
-- ich bin angespannt
-
-## intent:no_problem
-- danke
-- vielen dank
-- herzlichen dank
-- danke sehr
-- danke dir
-- dankeschön
-
-## intent:fast_start
-- ich möchte fasten
-- fasten start
-- starte zu fasten
-- fastenstart
-- fasten starten
-- fasten
-- start
-- ich will fasten
-- fastenbeginn
-- beginn fasten
-- start fasten
-
-## intent:fast_length_so_far
-- wie lange faste ich schon?
-- wann habe ich [16](fast_length) Stunden geschafft?
-- habe ich bald [20](fast_length) Stunden?
-- ist es bald vorbei?
-- faste ich schon?
-- bin ich schon am fasten?
-- habe ich es bald geschafft?
-- hab ichs bald hinter mir?
-- wie lange noch?
-
-## intent:fast_end
-- stop
-- stopp
-- fastenende
-- ende
-- fasten beenden
-- aufhören
-
-## intent:weight_add
-- ich habe mich gewogen
-- wiegen
-- waage
-- ich will mein gewicht angeben
-- ich möchte mein gewicht angeben
-- will gewicht eintragen
-- gewicht angeben
-- gewogen
-- gewicht eintragen
-- kg
-- kilogramm
-- kilos
-- kilo
-
-<!-- health data intents -->
-## intent:weight_value
-- ich wiege [66](weight_value)
-- mein gewicht ist [58](weight_value)
-- aktuell [74](weight_value)
-- [58](weight_value) kg
-- [84](weight_value) kilo
-- [92](weight_value) kg
-- [102](weight_value) kilogramm
-- [76](weight_value) kilogramm
-- [55](weight_value) KG
-- [113](weight_value) Kilogramm
-- ich habe [88](weight_value) kg gewogen
-- ich wiege [81](weight_value) kilo
-- [99](weight_value) kilo
-- [68](weight_value) kilos
-- ich wiege [74](weight_value) kilo
