+++ rasa/domain.yml
@@ -1,40 +1,7 @@
-actions:
-- action_end_fast
-- action_fasting_since
-- action_hello_world
-- action_start_fast
-- action_set_reminder_fast
-- action_forget_reminder_fast
-- action_entity_extract
-
-entities:
-- is_fasting
-- start_fast
-- hours_fasted
-- mins_fasted
-- weight_value
-- fast_length
-
-slots:
-  is_fasting:
-    initial_value: 0
-    type: categorical
-    values:
-    - '0'
-    - '1'
-  start_fast:
-    type: unfeaturized
-    initial_value: 'None' # rasa converts None to 'None' --> to prevent future errors, set 'None' right away
-  hours_fasted:
-    type: unfeaturized
-  mins_fasted:
-    type: unfeaturized
-  weight_value:
-    type: unfeaturized
-  requested_slot:
-    type: unfeaturized
-
-
+version: '2.0'
+session_config:
+  session_expiration_time: 0
+  carry_over_slots_to_new_session: true
 intents:
 - start
 - affirm
@@ -45,145 +12,143 @@ intents:
 - mood_great
 - mood_unhappy
 - no_problem
-# fast
 - fast_start
 - fast_length_so_far
 - fast_end
 - fast_reminder
-# health data
 - weight_add
 - weight_value
-
+entities:
+- is_fasting
+- start_fast
+- hours_fasted
+- mins_fasted
+- weight_value
+- fast_length
+slots:
+  is_fasting:
+    type: categorical
+    initial_value: 0
+    influence_conversation: true
+    values:
+    - '0'
+    - '1'
+  start_fast:
+    type: any
+    initial_value: None
+  hours_fasted:
+    type: any
+  mins_fasted:
+    type: any
+  weight_value:
+    type: any
+  requested_slot:
+    type: any
 responses:
-  ## conversational stmts
   utter_start:
-    - text: Hey! Ich helfe dir zu fasten und erfasse deine Fastenzeiten. Wenn du bereit bist, schreibe mir einfach, dass du fasten möchtest.
-
+  - text: Hey! Ich helfe dir zu fasten und erfasse deine Fastenzeiten. Wenn du bereit bist, schreibe mir einfach, dass du fasten möchtest.
   utter_goodbye:
-    - text: Ciao
-    - text: Bis bald
-    - text: Mach's gut
-  
+  - text: Ciao
+  - text: Bis bald
+  - text: Mach's gut
   utter_greet:
-    - text: Hey! Wie geht es dir? :)
-    - text: Hallo. Wie fühlst du dich?
-    - text: Hey, wie geht es dir?
-    - text: Hi. Wie fühlst du dich gerade?
-    - text: Hallo. Wie geht es dir heute?
-    - text: Guten Tag. Wie fühlst du dich?
-  
+  - text: Hey! Wie geht es dir? :)
+  - text: Hallo. Wie fühlst du dich?
+  - text: Hey, wie geht es dir?
+  - text: Hi. Wie fühlst du dich gerade?
+  - text: Hallo. Wie geht es dir heute?
+  - text: Guten Tag. Wie fühlst du dich?
   utter_happy:
-    - text: Super, weiter so!
-    - text: Das freut mich.
-    - text: Das klingt schön.
-    - text: Wunderbar
-  
+  - text: Super, weiter so!
+  - text: Das freut mich.
+  - text: Das klingt schön.
+  - text: Wunderbar
   utter_express_empathy:
-    - text: Das ist schade.
-    - text: Es scheint dir nicht gut zu gehen.
-    - text: Das tut mir leid zu hören.
-    - text: Du wirkst nachdenklich.
-
+  - text: Das ist schade.
+  - text: Es scheint dir nicht gut zu gehen.
+  - text: Das tut mir leid zu hören.
+  - text: Du wirkst nachdenklich.
   utter_I_am_fine:
-    - text: Mir geht es gut, danke.
-    - text: Ein bisschen langweilig ...
-    - text: Ich habe viel gelesen heute. Das hat mir gefallen.
-    - text: Ich fühle mich etwas nachdenklich heute.
-    - text: Ich glaube mir geht es so wie dir heute.
-  
+  - text: Mir geht es gut, danke.
+  - text: Ein bisschen langweilig ...
+  - text: Ich habe viel gelesen heute. Das hat mir gefallen.
+  - text: Ich fühle mich etwas nachdenklich heute.
+  - text: Ich glaube mir geht es so wie dir heute.
   utter_no_problem:
-    - text: kein Problem
-    - text: gern geschehen
-    - text: gerne
-  
+  - text: kein Problem
+  - text: gern geschehen
+  - text: gerne
   utter_ok:
-    - text: ok
-    - text: alles klar
-    - text: in Ordnung
-    - text: verstanden
-
+  - text: ok
+  - text: alles klar
+  - text: in Ordnung
+  - text: verstanden
   utter_noted:
-    - text: ok, ist notiert
-    - text: habe ich notiert
-    - text: alles klar, ich merke es mir
-    - text: ist eingetragen
-
+  - text: ok, ist notiert
+  - text: habe ich notiert
+  - text: alles klar, ich merke es mir
+  - text: ist eingetragen
   utter_thank_you:
-    - text: Danke, dass du mir das erzählt hast.
-    - text: Vielen Dank für das Teilen deiner Gedanken.
-    - text: Danke, dass du das mit mir geteilt hast.
-
-  ## fast stmts
+  - text: Danke, dass du mir das erzählt hast.
+  - text: Vielen Dank für das Teilen deiner Gedanken.
+  - text: Danke, dass du das mit mir geteilt hast.
   utter_ask_end_fast:
-    - buttons:
-      - payload: ja
-        title: ja
-      - payload: nein
-        title: nein
-      text: Du fastest seit {hours_fasted} und {mins_fasted}. Möchtest du das Fasten
-        jetzt beenden?
-  
+  - buttons:
+    - payload: ja
+      title: ja
+    - payload: nein
+      title: nein
+    text: Du fastest seit {hours_fasted} und {mins_fasted}. Möchtest du das Fasten jetzt beenden?
   utter_ask_fast:
-    - buttons:
-      - payload: ja
-        title: ja
-      - payload: nein
-        title: nein
-      text: Möchtest du fasten?
-  
+  - buttons:
+    - payload: ja
+      title: ja
+    - payload: nein
+      title: nein
+    text: Möchtest du fasten?
   utter_end_fast:
-    - text: Ich habe deine Fastenuhr gestoppt. Du hast {hours_fasted} und {mins_fasted}
-        gefastet. Glückwunsch :)
-
+  - text: Ich habe deine Fastenuhr gestoppt. Du hast {hours_fasted} und {mins_fasted} gefastet. Glückwunsch :)
   utter_info_fast:
-    - text: Du fastest seit {hours_fasted} und {mins_fasted}.
-  
+  - text: Du fastest seit {hours_fasted} und {mins_fasted}.
   utter_info_not_fasting_ask_fast:
-    - buttons:
-      - payload: ja
-        title: ja
-      - payload: nein
-        title: nein
-      text: Du fastest aktuell nicht. Möchtest du starten?
-    
+  - buttons:
+    - payload: ja
+      title: ja
+    - payload: nein
+      title: nein
+    text: Du fastest aktuell nicht. Möchtest du starten?
   utter_start_fast:
-    - text: Ich habe deine Fastenuhr gestartet. Frohes Fasten!
-    
+  - text: Ich habe deine Fastenuhr gestartet. Frohes Fasten!
   utter_reminder_fast:
-    - text: Du fastest seit {hours_fasted}! :)
-
-  ## health data stmts
+  - text: Du fastest seit {hours_fasted}! :)
   utter_ask_weight:
-    - buttons:
-      - payload: ja
-        title: ja
-      - payload: nein
-        title: nein
-      text: Möchtest du dein Gewicht eintragen?
-
+  - buttons:
+    - payload: ja
+      title: ja
+    - payload: nein
+      title: nein
+    text: Möchtest du dein Gewicht eintragen?
   utter_ask_weight_value:
-    - text: Wie viel hast du gewogen?
-
-  ## journal stmts
+  - text: Wie viel hast du gewogen?
   utter_ask_talk:
-    - buttons:
-      - payload: ja
-        title: ja
-      - payload: nein
-        title: nein
-      text: Möchtest du beschreiben, was dich beschäftigt?
-
+  - buttons:
+    - payload: ja
+      title: ja
+    - payload: nein
+      title: nein
+    text: Möchtest du beschreiben, was dich beschäftigt?
   utter_ask_feelings:
-    - text: Was beschäftigt dich?
-    - text: Was geht dir durch den Kopf?
-
+  - text: Was beschäftigt dich?
+  - text: Was geht dir durch den Kopf?
   utter_ask_countermeasures:
-    - text: Wie möchtest du mit deiner Situation umgehen?
-    - text: Siehst du eine Möglichkeit, deine Situation zu verändern?
-    - text: Gibt es etwas, das du tun könntest, um dich wohler zu fühlen?
-  
-
-session_config:
-  session_expiration_time: 0
-  carry_over_slots_to_new_session: true
-  
+  - text: Wie möchtest du mit deiner Situation umgehen?
+  - text: Siehst du eine Möglichkeit, deine Situation zu verändern?
+  - text: Gibt es etwas, das du tun könntest, um dich wohler zu fühlen?
+actions:
+- action_end_fast
+- action_fasting_since
+- action_hello_world
+- action_start_fast
+- action_set_reminder_fast
+- action_forget_reminder_fast
+- action_entity_extract
